refactor(auditing): 수정시간 변경 할 수 있는 클래스 구현 및 사용
