+++ backend/src/main/java/codezap/template/domain/Template.java
@@ -18,7 +18,7 @@ import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.DynamicUpdate;
 
 import codezap.category.domain.Category;
-import codezap.global.auditing.BaseTimeEntity;
+import codezap.global.auditing.SkipModifiedAtUpdateBaseTimeEntity;
 import codezap.member.domain.Member;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -32,7 +32,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Getter
 @EqualsAndHashCode(of = "id", callSuper = false)
-public class Template extends BaseTimeEntity {
+public class Template extends SkipModifiedAtUpdateBaseTimeEntity {
 
     private static final Long LIKES_COUNT_DEFAULT = 0L;
